pyproject.toml
[tool.poetry]
name = "fabric"
version = "0.1.0"
description = "Infectious Disease Modeling AI-Platform"
authors = ["Fitra Kacamarga <fitra@eureka.ai>"]

[tool.poetry.dependencies]
python = "^3.6.8"
prefect = "0.11.4"
pyspark = "2.3.0"
pymc3 = "3.8"
arviz = "^0.8.3"
pandas = "^1.0.4"
numpy = "^1.18.5"
scipy = "^1.4.1"
fastapi = "^0.56.1"
uvicorn = "^0.11.5"
streamlit = "^0.61.0"
pyyaml = "^5.3.1"
plotly = "^4.8.1"
pyarrow = "^0.17.1"
dask = {version = "^2.18.1", extras = ["dataframe"]}
findspark = "^1.4.2"
typer = "^0.2.1"
lmfit = "^1.0.1"
symfit = {git = "https://github.com/tBuLi/symfit"}
seirsplus = "^0.1.4"
poetry = "^1.0.9"
ansible = "^2.9.10"
supervisor = "^4.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
pylint = "^2.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
